LayoutTests/platform/ios-simulator/userscripts/user-script-plugin-document-expected.txt

CONSOLE MESSAGE: line 1: End Script: user-script-plugin-document.html
